### 2019-11-08 #1871 @whedon
[PRE REVIEW]: FLAM: Fast Linear Algebra in MATLAB
**Submitting author:** @klho (<a href="http://orcid.org/0000-0001-5450-4966">Kenneth L. Ho</a>)
**Repository:** <a href="https://github.com/klho/FLAM" target ="_blank">https://github.com/klho/FLAM</a>
**Version:** v1.1.0 (will update on completion)
