+++ paper/paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 
 # Summary
 
-Many large matrices in science and engineering possess a special hierarchical low-rank structure that enables fast multiplication and inversion, among other fundamental operations. Such matrices commonly occur in physical problems, including the classical integral (IE) and differential equations (DE) of potential theory, as well as, for example, in kernel methods for high-dimensional statistics and machine learning. Owing to their ubiquity and practical importance, a wide variety of techniques for exploiting this structure have been developed, appearing in the literature under an assortment of related names such as the fast multiple method (FMM) [@greengard:1987:j-comput-phys], $\mathcal{H}$- and $\mathcal{H}^2$-matrices [@hackbusch:1999:computing], hierarchically semiseparable matrices [@xia:2010:numer-linear-algebra-appl], hierarchically block separable matrices [@gillman:2012:front-math-china], and hierarchical off-diagonal low-rank matrices [@ambikasaran:2013:j-sci-comput]. Many of these now also have accompanying open-source software packages: ``FMMLIB`` [@gimbutas:2015:commun-comput-phys], ``H2Lib``, ``STRUMPACK`` [@rouet:2016:acm-trans-math-softw; @ghysels:2016:siam-j-sci-comput], ``HODLRlib`` [@ambikasaran:2019:j-open-source-softw], ``hm-toolbox`` [@massei:2019:arxiv], etc.
+Many large matrices in science and engineering possess a special hierarchical low-rank structure that enables fast multiplication and inversion, among other fundamental operations. Such matrices commonly occur in physical problems, including the classical integral (IE) and differential equations (DE) of potential theory, as well as, for example, in kernel methods for high-dimensional statistics and machine learning. Owing to their ubiquity and practical importance, a wide variety of techniques for exploiting this structure have been developed, appearing in the literature under an assortment of related names such as the fast multiple method (FMM) [@greengard:1987:j-comput-phys], $\mathcal{H}$- and $\mathcal{H}^2$-matrices [@hackbusch:1999:computing], hierarchically semiseparable matrices [@xia:2010:numer-linear-algebra-appl], hierarchically block separable matrices [@gillman:2012:front-math-china], and hierarchical off-diagonal low-rank matrices [@ambikasaran:2013:j-sci-comput]. Many of these now also have accompanying open-source software packages: ``FMMLIB`` [@gimbutas:2015:commun-comput-phys], ``H2Lib`` (http://www.h2lib.org), ``STRUMPACK`` [@rouet:2016:acm-trans-math-softw; @ghysels:2016:siam-j-sci-comput], ``HODLRlib`` [@ambikasaran:2019:j-open-source-softw], ``hm-toolbox`` [@massei:2019:arxiv], etc.
 
 ``FLAM`` is a MATLAB (and Octave-compatible) library in this same vein, implementing methods mostly in the "recursive skeletonization" style. Briefly, the core algorithms take as input a matrix implicitly defined by a function to generate arbitrary entries and whose rows/columns are associated with points in $\mathbb{R}^d$. The induced geometry exposes the low-rank matrix blocks, which are then compressed or sparsified using the interpolative decomposition (ID) in a multilevel manner. Being written in MATLAB, the code is quite readable and easy to extend, especially for research purposes, though it may not be the most performant. Still, it is reasonably feature-complete; currently provided algorithms include, for dense IE-like matrices with dimensions $M \geq N$:
 
