### 2019-11-13 f01a2fc
Correct gqgw doc.
