### 2019-11-13 #1871 @melissawm
@pitsianis would you be available to review this software paper for JOSS? Thank you
