### 2019-11-14 #1871 @sivaramambikasaran
> @sivaramambikasaran, would you be willing to review this software paper for JOSS? Thank you!

@melissawm Sure would ❤ to. The only thing is I can only start next week since we now have the end sof the semester craze going on.
