+++ core/chksymm.m
@@ -7,8 +7,7 @@
 
 function symm = chksymm(symm)
   symm = lower(symm);
-  assert(strcmp(symm,'n') || strcmp(symm,'s') || ...
-         strcmp(symm,'h') || strcmp(symm,'p'), ...
-         'FLAM:chksymm:invalidSymm', ...
+  assert(strcmp(symm,'n') || strcmp(symm,'s') || strcmp(symm,'h') || ...
+         strcmp(symm,'p'),'FLAM:chksymm:invalidSymm', ...
          'Symmetry parameter must be one of ''N'', ''S'', ''H'', or ''P''.')
 end
