+++ ifmm/ifmm.m
@@ -98,9 +98,8 @@ function F = ifmm(A,rx,cx,occ,rank_or_tol,pxyfun,opts)
 
   % check inputs
   opts.store = lower(opts.store);
-  assert(strcmp(opts.store,'n') || strcmp(opts.store,'s') || ...
-         strcmp(opts.store,'r') || strcmp(opts.store,'a'), ...
-         'FLAM:ifmm:invalidStore', ...
+  assert(opts.store == 'n' || opts.store == 's' || opts.store == 'r' || ...
+         opts.store == 'a','FLAM:ifmm:invalidStore', ...
          'Storage parameter must be one of ''N'', ''S'', ''R'', or ''A''.')
   opts.symm = chksymm(opts.symm);
   if opts.symm == 'p', opts.symm = 'h'; end
