### 2019-11-26 c856b00
Minor bug fix in trisphere_subdiv. Use vertex mode in proxies.
