+++ rskelf/test/ie_cube1.m
@@ -17,7 +17,7 @@ function ie_cube1(n,occ,p,rank_or_tol,symm,doiter)
   clear x1 x2 x3
   N = size(x,2);
   % proxy points are quasi-uniform sampling of scaled 1.5-radius sphere
-  proxy = trisphere_subdiv(p); r = randperm(size(proxy,2));
+  proxy = trisphere_subdiv(p,'v'); r = randperm(size(proxy,2));
   proxy = proxy(:,r(1:p));  % reference proxy points are for unit box [-1, 1]^3
 
   % compute diagonal quadratures
