+++ ifmm/test/mv_sphere2.m
@@ -21,7 +21,7 @@ function mv_sphere2(m,n,k,occ,p,rank_or_tol,near,store)
   M = size(rx,2);
   N = size(cx,2);
   % proxy points are quasi-uniform sampling of scaled 1.5-radius sphere
-  proxy = trisphere_subdiv(p); r = randperm(size(proxy,2));
+  proxy = trisphere_subdiv(p,'v'); r = randperm(size(proxy,2));
   proxy = proxy(:,r(1:p));  % reference proxy points are for unit box [-1, 1]^3
 
   % compress matrix
