+++ geom/trisphere_subdiv.m
@@ -4,8 +4,21 @@
 %    with vertices V and faces F of size at least N by recursively subdividing a
 %    base icosahedron triangulation. The vertices of triangle I are V(:,F(:,I)).
 %    The number of triangles on output can be 20, 80, 320, 1280, etc.
+%
+%    [V,F] = TRISPHERE_SUBDIV(N,VFMIN) produces a triangulation of size at least
+%    N, where size is measured in terms of vertices if VFMIN = 'V' and of faces
+%    if VFMIN = 'F'.
 
-function [V,F] = trisphere_subdiv(n)
+function [V,F] = trisphere_subdiv(n,vfmin)
+
+  % set default parameters
+  if nargin < 2 || isempty(vfmin), vfmin = 'f'; end
+
+  % check inputs
+  vfmin = lower(vfmin);
+  assert(strcmp(vfmin,'v') || strcmp(vfmin,'f'), ...
+         'FLAM:trisphere_subdiv:invalidVfmin', ...
+         'Minimum vertex/face parameter must be one of ''V'' or ''F''.')
 
   % base triangulation
   t = (1 + sqrt(5))/2;
@@ -42,7 +55,9 @@ function [V,F] = trisphere_subdiv(n)
         7  3 11
         9  7  8
        10  9  2];
-  m = size(F,1);
+  if vfmin == 'v', m = size(V,1);
+  else,            m = size(F,1);
+  end
 
   % recursively subdivide and project
   while m < n
@@ -52,15 +67,18 @@ function [V,F] = trisphere_subdiv(n)
     M = M./(sqrt(sum(M.^2,2))*ones(1,3));
     [M,~,I] = unique(M,'rows');
     nv = size(V,1);
-    F12 = nv + I(    1:  m);
-    F13 = nv + I(  m+1:2*m);
-    F23 = nv + I(2*m+1:3*m);
+    nf = size(F,1);
+    F12 = nv + I(     1:  nf);
+    F13 = nv + I(  nf+1:2*nf);
+    F23 = nv + I(2*nf+1:3*nf);
     V = [V; M];
     F = [F(:,1)  F12    F13
           F12   F(:,2)  F23
           F13    F23   F(:,3)
           F12    F23    F13];
-    m = size(F,1);
+    if vfmin == 'v', m = size(V,1);
+    else,            m = size(F,1);
+    end
   end
 
   % rotate outputs
