+++ ifmm/test/mv_cube1.m
@@ -25,7 +25,7 @@ function mv_cube1(m,n,occ,p,rank_or_tol,near,store)
   rx = rand(3,m); M = size(rx,2);  % row points
   cx = rand(3,n); N = size(cx,2);  % col points
   % proxy points are quasi-uniform sampling of scaled 1.5-radius sphere
-  proxy = trisphere_subdiv(p); r = randperm(size(proxy,2));
+  proxy = trisphere_subdiv(p,'v'); r = randperm(size(proxy,2));
   proxy = proxy(:,r(1:p));  % reference proxy points are for unit box [-1, 1]^3
 
   % compress matrix
