### 2019-11-29 #1906 @sivaramambikasaran
@melissawm @kyleniemeyer Wanted to check if MATLAB code is okay under JOSS guidelines given that the underlying ecosystem for the code is neither free nor open source? (My MATLAB license expired today and prompted this question).
