### 2019-11-29 #1906 @pitsianis
I also agree that there is no issue with the use of MATLAB, even if it were not compatible with Octave (I did not verify such a claim).

The issue I have is with the name of the project, because the title/acronym refers to a much more general subject and as an acronym it is very close to this project

https://www.cs.utexas.edu/~flame

The submitter should select a title and acronym that reflects more accurately the capabilities of this software.
