### 2019-12-03 099038d
Simplify; doesn't seem necessary.
