+++ mf/mf_sv.m
@@ -1,6 +1,6 @@
 % MF_SV  Solve using multifrontal factorization.
 %
-%    Typical complexity: same as RSKELF_MV.
+%    Typical complexity: same as MF_MV.
 %
 %    Y = MF_SV(F,X) produces the matrix Y by applying the inverse of the
 %    factored matrix F to the matrix X.
