### 2019-12-10 88d8e27
Simplify some tests. Use quasi-uniform sampling for certain least squares tests. More comments.
