+++ ifmm/test/mv_square1.m
@@ -10,11 +10,11 @@
 %   - check multiply error/time
 %   - check adjoint multiply error/time
 
-function mv_square1(m,n,occ,p,rank_or_tol,near,store)
+function mv_square1(M,N,occ,p,rank_or_tol,near,store)
 
   % set default parameters
-  if nargin < 1 || isempty(m), m = 16384; end  % number of row points
-  if nargin < 2 || isempty(n), n =  8192; end  % number of col points
+  if nargin < 1 || isempty(M), M = 16384; end  % number of row points
+  if nargin < 2 || isempty(N), N =  8192; end  % number of col points
   if nargin < 3 || isempty(occ), occ = 128; end
   if nargin < 4 || isempty(p), p = 64; end  % number of proxy points
   if nargin < 5 || isempty(rank_or_tol), rank_or_tol = 1e-9; end
@@ -22,8 +22,8 @@ function mv_square1(m,n,occ,p,rank_or_tol,near,store)
   if nargin < 7 || isempty(store), store = 'n'; end  % no storage
 
   % initialize
-  rx = rand(2,m); M = size(rx,2);  % row points
-  cx = rand(2,n); N = size(cx,2);  % col points
+  rx = rand(2,M);                                              % row points
+  cx = rand(2,N);                                              % col points
   theta = (1:p)*2*pi/p; proxy = 1.5*[cos(theta); sin(theta)];  % proxy points
   % reference proxy points are for unit box [-1, 1]^2
 
