### 2019-12-12 11035a6
Rename rskel/test/ols_square -> ols_square1; add ols_square2 for regularized Laplace fitting.
