+++ rskel/test/ols_square1.m
@@ -1,0 +1,174 @@
+% Overdetermined least squares on the unit square, thin-plate splines.
+%
+% This is the direct analogue of OLS_LINE for quasi-uniform points in 2D.
+
+function ols_square1(M,N,lambda,occ,p,rank_or_tol,store,doiter)
+
+  % set default parameters
+  if nargin < 1 || isempty(M), M = 16384; end  % number of row points
+  if nargin < 2 || isempty(N), N =  8192; end  % number of col points
+  if nargin < 3 || isempty(lambda), lambda = 0.1; end  % regularization
+  if nargin < 4 || isempty(occ), occ = 128; end
+  if nargin < 5 || isempty(p), p = 64; end  % half number of proxy points
+  if nargin < 6 || isempty(rank_or_tol), rank_or_tol = 1e-6; end
+  if nargin < 7 || isempty(store), store = 'a'; end  % FMM storage mode
+  if nargin < 8 || isempty(doiter), doiter = 1; end  % naive LSQR/CG?
+
+  % initialize
+  m = ceil(sqrt(M)); [x1,x2] = ndgrid((1:m)/m); rx = [x1(:) x2(:)]';
+  r = randperm(size(rx,2)); rx = rx(:,r(1:M));  % row points
+  n = ceil(sqrt(N)); [x1,x2] = ndgrid((1:n)/n); cx = [x1(:) x2(:)]';
+  r = randperm(size(cx,2)); cx = cx(:,r(1:N));  % col points
+  clear x1 x2
+  % proxy points -- two concentric rings (thin-plate splines are Green's
+  % function for biharmonic equation; no Green's theorem)
+  theta = (1:p)*2*pi/p;
+  proxy = 1.5*[cos(theta); sin(theta)]; proxy = [proxy 2*proxy];
+  % reference proxy points are for unit box [-1, 1]^2
+
+  % compress matrix using RSKEL
+  Afun = @(i,j)Afun_(i,j,rx,cx);
+  pxyfun = @(rc,rx,cx,slf,nbr,l,ctr)pxyfun_(rc,rx,cx,slf,nbr,l,ctr,proxy);
+  opts = struct('verb',1);
+  tic; F = rskel(Afun,rx,cx,occ,rank_or_tol,pxyfun,opts); t = toc;
+  w = whos('F'); mem = w.bytes/1e6;
+  fprintf('rskel time/mem: %10.4e (s) / %6.2f (MB)\n',t,mem)
+
+  % compress matrix using IFMM
+  rank_or_tol = max(rank_or_tol*1e-2,1e-15);  % higher accuracy for reference
+  opts = struct('store',store);
+  tic; G = ifmm(Afun,rx,cx,occ,rank_or_tol,pxyfun,opts); t = toc;
+  w = whos('G'); mem = w.bytes/1e6;
+  fprintf('ifmm time/mem: %10.4e (s) / %6.2f (MB)\n',t,mem)
+
+  % test accuracy using randomized power method
+  X = rand(N,1);
+  X = X/norm(X);
+
+  % NORM(A - F)/NORM(A)
+  tic; rskel_mv(F,X); t = toc;  % for timing
+  err = snorm(N,@(x)(ifmm_mv(G,x,Afun,'n') - rskel_mv(F,x,'n')), ...
+                @(x)(ifmm_mv(G,x,Afun,'c') - rskel_mv(F,x,'c')));
+  err = err/snorm(N,@(x)ifmm_mv(G,x,Afun,'n'),@(x)ifmm_mv(G,x,Afun,'c'));
[…]
