### 2019-12-13 b1bdf84
Simplify slightly.
