### 2019-12-16 84a4189
Use column-oriented layout for ctr data -- not sure why I didn't have this way before...
