+++ ifmm/test/mv_square2.m
@@ -65,7 +65,7 @@ end
 
 % proxy function
 function [Kpxy,nbr] = pxyfun_(rc,rx,cx,slf,nbr,l,ctr,proxy,k)
-  pxy = proxy*l + ctr';  % scale and translate reference points
+  pxy = proxy*l + ctr;  % scale and translate reference points
   if rc == 'r'
     Kpxy = Kfun(rx(:,slf),pxy,k);
     dx = cx(1,nbr) - ctr(1);
