+++ rskel/test/uls_square.m
@@ -130,7 +130,7 @@ end
 
 % proxy function
 function [Kpxy,nbr] = pxyfun_(rc,rx,cx,slf,nbr,l,ctr,proxy)
-  pxy = proxy*l + ctr';  % scale and translate reference points
+  pxy = proxy*l + ctr;  % scale and translate reference points
   if rc == 'r'
     Kpxy = Kfun(rx(:,slf),pxy);
     dx = cx(1,nbr) - ctr(1);
