+++ hifie/base/hifie3_base.m
@@ -82,35 +82,35 @@ function F = hifie3_base(A,x,occ,rank_or_tol,idfun,pxyfun,opts)
 
         % generate face centers
         if d == 2
-          ctr = zeros(6*nbox,3);
+          ctr = zeros(3,6*nbox);
           box2ctr = cell(nbox,1);
           for i = t.lvp(lvl)+1:t.lvp(lvl+1)  % cells on this level
             j = i - t.lvp(lvl);              % local cell index
             idx = 6*(j-1)+1:6*j;             % face indices
             off = [0 0 -1; 0 -1 0; -1 0 0;   % offset from cell center
-                   0 0  1; 0  1 0;  1 0 0];
-            ctr(idx,:) = t.nodes(i).ctr + 0.5*l*off;  % face centers
+                   0 0  1; 0  1 0;  1 0 0]';
+            ctr(:,idx) = t.nodes(i).ctr + 0.5*l*off;  % face centers
             box2ctr{j} = idx;  % mapping from each cell to its faces
           end
 
         % generate edge centers
         elseif d == 1
-          ctr = zeros(12*nbox,3);
+          ctr = zeros(3,12*nbox);
           box2ctr = cell(nbox,1);
           for i = t.lvp(lvl)+1:t.lvp(lvl+1)  % cells on this level
             j = i - t.lvp(lvl);              % local cell index
             idx = 12*(j-1)+1:12*j;           % edge indices
             off = [ 0 -1 -1;  0 -1 1; 0  1 -1; 0 1 1;  % offset from cell center
                    -1  0 -1; -1  0 1; 1  0 -1; 1 0 1;
-                   -1 -1  0; -1  1 0; 1 -1  0; 1 1 0];
-            ctr(idx,:) = t.nodes(i).ctr + 0.5*l*off;  % edge centers
+                   -1 -1  0; -1  1 0; 1 -1  0; 1 1 0]';
+            ctr(:,idx) = t.nodes(i).ctr + 0.5*l*off;  % edge centers
             box2ctr{j} = idx;  % mapping from each cell to its edges
           end
         end
 
         % restrict to unique shared centers
         idx = round(2*(ctr - t.nodes(1).ctr)/l);  % displacement from root
-        [~,i,j] = unique(idx,'rows');             % unique indices
+        [~,i,j] = unique(idx','rows');            % unique indices
         p = find(histc(j,1:max(j)) > 1);          % shared indices
         % for edges, further restrict to those shared diagonally across boxes
         if d == 1
@@ -120,33 +120,33 @@ function F = hifie3_base(A,x,occ,rank_or_tol,idfun,pxyfun,opts)
           [~,q] = sort(j); k = k(q);  % sort by mapping to shared centers
           q = [0; find(diff(j(q))); length(j)];  % index array to shared centers
           for ip = 1:np               % loop over shared centers
+            % centers of corresponding parent boxes
             c = [t.nodes(t.lvp(lvl)+k(q(p(ip))+1:q(p(ip)+1))).ctr];
-            c = reshape(c,3,[])';     % centers of corresponding parent boxes
-            dx = c(:,1) - c(:,1)';    % distance between boxes
-            dy = c(:,2) - c(:,2)';
-            dz = c(:,3) - c(:,3)';
+            dx = c(1,:)' - c(1,:);    % distance between boxes
+            dy = c(2,:)' - c(2,:);
+            dz = c(3,:)' - c(3,:);
             dist = round((abs(dx) + abs(dy) + abs(dz))/l);  % convert to integer
[…]
