+++ hifie/test/cov_square2.m
@@ -136,7 +136,7 @@ end
 
 % proxy function
 function [Kpxy,nbr] = pxyfun_(x,slf,nbr,l,ctr,proxy,scale)
-  pxy = proxy*l + ctr';  % scale and translate reference points
+  pxy = proxy*l + ctr;  % scale and translate reference points
   Kpxy = Kfun(pxy,x(:,slf),scale);
   dx = x(1,nbr) - ctr(1);
   dy = x(2,nbr) - ctr(2);

[…]
