### 2019-12-19 a81ad24
Allow anisotropic domain bisection (though still with "mostly" hypercube nodes); otherwise, e.g., rskelf/test/ie_ellipse just doesn't work for high aspect ratio.
