+++ rskel/test/ie_square2.m
@@ -100,23 +100,20 @@ end
 
 % proxy function
 function [Kpxy,nbr] = pxyfun_(rc,rx,cx,slf,nbr,l,ctr,proxy)
-  pxy = proxy*l + ctr;  % scale and translate reference points
+  pxy = proxy.*l + ctr;  % scale and translate reference points
   % proxy interaction is kernel evaluation between proxy points and row/column
   % points being compressed, scaled to match the matrix scale
   N = size(rx,2);
   if rc == 'r'
     Kpxy = Kfun(rx(:,slf),pxy)/N;
-    dx = cx(1,nbr) - ctr(1);
-    dy = cx(2,nbr) - ctr(2);
+    dr = cx(:,nbr) - ctr;
   else
     Kpxy = Kfun(pxy,cx(:,slf))/N;
-    dx = rx(1,nbr) - ctr(1);
-    dy = rx(2,nbr) - ctr(2);
+    dr = rx(:,nbr) - ctr;
   end
-  % proxy points form circle of scaled radius 1.5 around current box
-  % keep among neighbors only those within circle
-  dist = sqrt(dx.^2 + dy.^2);
-  nbr = nbr(dist/l < 1.5);
+  % proxy points form ellipse of scaled "radius" 1.5 around current box
+  % keep among neighbors only those within ellipse
+  nbr = nbr(sum((dr./l).^2) < 1.5^2);
 end
 
 % FFT multiplication
