+++ rskel/test/mv_line.m
@@ -65,7 +65,7 @@ end
 
 % proxy function
 function [Kpxy,nbr] = pxyfun_(rc,rx,cx,slf,nbr,l,ctr,proxy)
-  pxy = proxy*l + ctr;  % scale and translate reference points
+  pxy = proxy.*l + ctr;  % scale and translate reference points
   if rc == 'r'
     Kpxy = Kfun(rx(:,slf),pxy);
     dr = cx(nbr) - ctr;
@@ -75,6 +75,5 @@ function [Kpxy,nbr] = pxyfun_(rc,rx,cx,slf,nbr,l,ctr,proxy)
   end
   % proxy points form interval of scaled radius 1.5 around current box
   % keep among neighbors only those within interval
-  dist = abs(dr);
-  nbr = nbr(dist/l < 1.5);
+  nbr = nbr(abs(dr)/l < 1.5);
 end
