+++ rskel/rskel.m
@@ -156,7 +156,7 @@ function F = rskel(A,rx,cx,occ,rank_or_tol,pxyfun,opts)
     ts = tic;
     nlvl = nlvl + 1;
     nrrem1 = sum(rrem); ncrem1 = sum(crem);  % remaining row/cols at start
-    l = t.lrt/2^(lvl - 1);
+    l = t.l(:,lvl);
 
     % pull up skeletons from children
     for i = t.lvp(lvl)+1:t.lvp(lvl+1)
