### 2019-12-19 8027376
Uniformly use restart = 32 in GMRES tests to avoid killing memory.
