+++ hifie/test/ie_sphere1.m
@@ -180,7 +180,7 @@ end
 
 % proxy function
 function [Kpxy,nbr] = pxyfun_(x,slf,nbr,l,ctr,proxy,nu,area)
-  pxy = proxy*l + ctr;  % scale and translate reference points
+  pxy = proxy.*l + ctr;  % scale and translate reference points
   % proxy interaction is kernel evaluation between proxy points and row/column
   % points being compressed, scaled to match the matrix scale
   N = size(x,2);  % from proxy points to centroids: use average triangle area
@@ -188,34 +188,27 @@ function [Kpxy,nbr] = pxyfun_(x,slf,nbr,l,ctr,proxy,nu,area)
   % from triangles to proxy points: use actual triangle areas
   Kpxy_c = Kfun(pxy,x(:,slf),'d',nu(:,slf)).*area(slf);
   Kpxy = [Kpxy_c; Kpxy_r'];  % stack row/column proxy matrices
-  dx = x(1,nbr) - ctr(1);
-  dy = x(2,nbr) - ctr(2);
-  dz = x(3,nbr) - ctr(3);
-  % proxy points form sphere of scaled radius 1.5 around current box
-  % keep among neighbors only those within sphere
-  dist = sqrt(dx.^2 + dy.^2 + dz.^2);
-  nbr = nbr(dist/l < 1.5);
+  % proxy points form ellipsoid of scaled "radius" 1.5 around current box
+  % keep among neighbors only those within ellipsoid
+  nbr = nbr(sum(((x(:,nbr) - ctr)./l).^2) < 1.5^2);
 end
 
 % proxy function for IFMM
 function [Kpxy,nbr] = pxyfun_ifmm_(rc,rx,cx,slf,nbr,l,ctr,proxy,nu,area)
-  pxy = proxy*l + ctr;  % scale and translate reference points
+  pxy = proxy.*l + ctr;  % scale and translate reference points
   % proxy interaction is kernel evaluation between proxy points and row/column
   % points being compressed, scaled to match the matrix scale
   if rc == 'r'
     % from proxy points to centroids: use average triangle area
     N = size(rx,2);
     Kpxy = Kfun(rx(:,slf),pxy,'s')*(4*pi/N);
-    dx = cx(1,nbr) - ctr(1);
-    dy = cx(2,nbr) - ctr(2);
+    dr = cx(:,nbr) - ctr;
   else
     % from triangles to proxy points: use actual triangle areas
     Kpxy = Kfun(pxy,cx(:,slf),'d',nu(:,slf)).*area(slf);
-    dx = rx(1,nbr) - ctr(1);
-    dy = rx(2,nbr) - ctr(2);
+    dr = rx(:,nbr) - ctr;
   end
-  % proxy points form sphere of scaled radius 1.5 around current box
-  % keep among neighbors only those within sphere
-  dist = sqrt(dx.^2 + dy.^2);
-  nbr = nbr(dist/l < 1.5);
+  % proxy points form ellipsoid of scaled "radius" 1.5 around current box
+  % keep among neighbors only those within ellipsoid
+  nbr = nbr(sum((dr./l).^2) < 1.5^2);
 end
