### 2019-12-22 d1610fd
Rename rskel/test/uls_square -> uls_square1; add uls_square2. Minor fix for ols_square2.
