+++ paper/paper.bib
@@ -275,12 +275,15 @@
   doi = {10.1145/2930660}
 }
 
-@article{tian:2019:arxiv,
+@article{tian:2019:commun-math-sci,
   author = {Tian, X. and Engquist, B.},
   title = {{Fast algorithm for computing nonlocal operators with finite interaction distance}},
-  journal = {Preprint},
-  pages = {arXiv:1905.08375 [math.NA]},
-  year = {2019}
+  journal = {Commun. Math. Sci.},
+  volume = {17},
+  number = {6},
+  pages = {1653-1670},
+  year = {2019},
+  doi = {10.4310/CMS.2019.v17.n6.a7}
 }
 
 @phdthesis{wang:2016:new-jersey-inst-tech,
