+++ paper/paper.md
@@ -42,7 +42,7 @@ Similarly, for sparse DE-like matrices, we have:
 
 Most of these algorithms have previously been published though some are perhaps new (if but straightforward modifications or extensions of existing ones).
 
-``FLAM`` was originally created to support the computations in @ho:2016a:comm-pure-appl-math and @ho:2016b:comm-pure-appl-math, and has since been used in @liu:2016:dartmouth-coll, @wang:2016:new-jersey-inst-tech, @corona:2017:j-comput-phys, @fang:2017:j-fluid-mech, @jiang:2017:multiscale-model-simul, @minden:2017b:multiscale-model-simul, @fan:2019:j-comput-phys, @li:2019:eng-anal-bound-elem, @wang:2019:j-sci-comput, @askham:2019:arxiv, and @tian:2019:arxiv. It has also served as the starting point for various more sophisticated codes such as those described in @minden:2016:multiscale-model-simul, @li:2017:res-math-sci, @minden:2017:stanford-univ, @minden:2017a:multiscale-model-simul, and @feliu-faba:2018:arxiv.
+``FLAM`` was originally created to support the computations in @ho:2016a:comm-pure-appl-math and @ho:2016b:comm-pure-appl-math, and has since been used in @liu:2016:dartmouth-coll, @wang:2016:new-jersey-inst-tech, @corona:2017:j-comput-phys, @fang:2017:j-fluid-mech, @jiang:2017:multiscale-model-simul, @minden:2017b:multiscale-model-simul, @fan:2019:j-comput-phys, @li:2019:eng-anal-bound-elem, @tian:2019:commun-math-sci, @wang:2019:j-sci-comput, and @askham:2019:arxiv. It has also served as the starting point for various more sophisticated codes such as those described in @minden:2016:multiscale-model-simul, @li:2017:res-math-sci, @minden:2017:stanford-univ, @minden:2017a:multiscale-model-simul, and @feliu-faba:2018:arxiv.
 
 We hope that ``FLAM`` will be a valuable research tool for the broader scientific community to explore the use of fast matrix methods in their applications of interest as well as to prototype new algorithmic ideas and implementations.
 
