### 2020-01-04 ef7bc54
Use more generic ID row-reduction if too tall-and-skinny; can use in hifie since now no randomization.
