+++ hifie/base/hifie_id.m
@@ -1,5 +1,5 @@
 % HIFIE_ID  Compression function for HIFIE2 and HIFIE3.
 
 function [sk,rd,T] = hifie_id(K,K1,K2,rank_or_tol)
-  [sk,rd,T] = id(K,rank_or_tol,0);  % no randomization for better rank control
+  [sk,rd,T] = id(K,rank_or_tol);
 end
