### 2020-02-04 a33f7f0
Fix some auxiliary HIFDE complexity estimates.
