### 2020-02-06 2ab8a41
Messed up hifde_spdiag estimate still.
