### 2020-02-08 klho/FLAM#5 @pitsianis
Title and acronym of this project (and of the corresponding JOSS paper) 
I strongly disagree with the title and acronym of this project. The title/acronym refers to a much more general subject:

There are many ways to do "Fast Linear Algebra"

 1) when the matrix A(i,j) is the evaluation of a kernel function K(r(i,j)) where r(i,j) is the distance between two points x(i) and x(j) from a low dimensional space (1, 2, or 3d for example) and the kernel K tends to zero when the distances increase.
 2) when the matrix is sparse
 3) when the matrix is banded
 4) when the matrix is low rank
 5) when the matrix is a Kronecker product
 6) when the matrix is diagonalizable by a fast transform matrix (Fourier, Wavelet, Hadamard, Haar etc)
 7) many more

Or when the matrix can be
a) reordered to,
b) extended to,
c) split into sub matrices,
d) factored into,
e) written as a summation

of combinations of some of the previous matrix types (1-6) and ways (a-e)

This project certainly does use some of the above, but it cannot claim the "Fast Linear Algebra" without some specification.

Also, as an acronym it is very close to a dense linear algebra code generation project at UT at Austin that has been around since 2000 and continues to be relevant. (Ken, Jack Poulson did his MS thesis with this group!)

[Formal Linear Algebra Methods Environment (FLAME)](https://www.cs.utexas.edu/~flame)

The submitter should select a title and acronym that reflects more accurately the capabilities of this software.
