### 2020-02-08 #1906 @pitsianis
I have opened this [issue](https://github.com/klho/FLAM/issues/5) regarding the title and acronym of this project.
