### 2020-03-04 klho/FLAM#5 @klho
There are certainly some fair points here, but given that this project has already been used in numerous other projects and publications, I would be extremely hesitant to change the name. I think that the most important concern is whether it is unambiguous from the README or the paper what the true scope of the project is -- of course, I think this has been achieved, but that is ultimately for the reviewers to decide.

As a compromise, I am not opposed to changing the title of the JOSS paper to something like "FLAM: Fast Linear Algebra in MATLAB - Algorithms for Hierarchical Matrices" -- provided that this does not violate some consistency requirement based on JOSS rules.

However, I would really prefer to see the full scope of all reviews before taking any such action in order to properly assess the full situation. So, with your understanding, I would like to delay any action here until then.
