### 2020-03-06 #1906 @pitsianis
> Thanks, @pitsianis . Are there any other concerns?

Do not get me wrong this is exceptional work, it solves a very complex and important problem successfully with simple and elegant code, it has been published in many good journal articles and this software deserves publishing at JOSS.

Now on my concerns. The claim is overly wide in the title. The first paragraph does put things in their right context. A better title will alleviate all my concerns on this.

The remaining issues are minor.

1. I would prefer to see a specific simple example to be presented so that a user not familiar with this work could appreciate the capabilities of this software. A single phrase description like "This example solves the Poisson equation on the unit square with Dirichlet boundary conditions" won't say much to the great majority of a diverse, non-specialist audience, right? But if, for example, show in addition the discretized matrix, perhaps more people might link it with something they have seen before.

2. MATLAB users expect when issuing `help command` to see the list of arguments that `command` expects and a description of what they are. You have done a great job documenting the all major functions, but none of the */test functions.

3. State in the paper the maximum number of dimensions that points belong to in $\mathbb{R}^d$, that the software supports. In several locations there are mentions like for $d > 1$, for all $d$ etc but could I try it for $d = 1000$? ;-)

4. No explicit statement in the article whether there exists related software and how it compares to this one.

With the above fixed, I will conclude my (very delayed, sorry) review accepting this work.
