### 2020-03-06 #1906 @klho
These seem very fair -- will get to them and let you know when I'm done.

Also, let me know what you think about modifying the paper (and README?) title as in [this comment](https://github.com/klho/FLAM/issues/5#issuecomment-594641864).
