### 2020-03-16 e23fe24
Document inputs for rskel/test.
