+++ rskelf/test/ie_square1.m
@@ -12,16 +12,25 @@
 %   - factor the matrix
 %   - check multiply/solve error/time
 %   - compare GMRES with/without preconditioning by approximate solve
+%
+% Inputs (defaults are used if not provided or set empty):
+%
+%   - N: number of discretization points in each dimension (default: N = 128)
+%   - OCC: tree occupancy parameter (default: OCC = 64)
+%   - P: number of proxy points (default: P = 64)
+%   - RANK_OR_TOL: local precision parameter (default: RANK_OR_TOL = 1e-6)
+%   - SYMM: symmetry parameter (default: SYMM = 'H')
+%   - DOITER: whether to run unpreconditioned GMRES (default: DOITER = 1)
 
 function ie_square1(n,occ,p,rank_or_tol,symm,doiter)
 
   % set default parameters
-  if nargin < 1 || isempty(n), n = 128; end  % number of points in each dim
+  if nargin < 1 || isempty(n), n = 128; end
   if nargin < 2 || isempty(occ), occ = 64; end
-  if nargin < 3 || isempty(p), p = 64; end  % number of proxy points
+  if nargin < 3 || isempty(p), p = 64; end
   if nargin < 4 || isempty(rank_or_tol), rank_or_tol = 1e-6; end
-  if nargin < 5 || isempty(symm), symm = 'h'; end  % symmetric/Hermitian
-  if nargin < 6 || isempty(doiter), doiter = 1; end  % unpreconditioned GMRES?
+  if nargin < 5 || isempty(symm), symm = 'h'; end
+  if nargin < 6 || isempty(doiter), doiter = 1; end
 
   % initialize
   [x1,x2] = ndgrid((1:n)/n); x = [x1(:) x2(:)]'; clear x1 x2;  % grid points
