+++ rskelf/test/fd_cube.m
@@ -1,15 +1,23 @@
 % Seven-point stencil on the unit cube, Poisson equation.
 %
 % This is basically the 3D analogue of FD_SQUARE.
+%
+% Inputs (defaults are used if not provided or set empty):
+%
+%   - N: number of discretization points in each dimension (default: N = 32)
+%   - OCC: tree occupancy parameter (default: OCC = 64)
+%   - RANK_OR_TOL: local precision parameter (default: RANK_OR_TOL = 1e-6)
+%   - SYMM: symmetry parameter (default: SYMM = 'P')
+%   - DOITER: whether to run unpreconditioned CG (default: DOITER = 1)
 
 function fd_cube(n,occ,rank_or_tol,skip,symm,doiter)
 
   % set default parameters
-  if nargin < 1 || isempty(n), n = 32; end  % number of points in each dimension
+  if nargin < 1 || isempty(n), n = 32; end
   if nargin < 2 || isempty(occ), occ = 64; end
   if nargin < 3 || isempty(rank_or_tol), rank_or_tol = 1e-6; end
-  if nargin < 4 || isempty(symm), symm = 'p'; end  % positive definite
-  if nargin < 5 || isempty(doiter), doiter = 1; end  % unpreconditioned CG?
+  if nargin < 4 || isempty(symm), symm = 'p'; end
+  if nargin < 5 || isempty(doiter), doiter = 1; end
 
   % initialize
   [x1,x2,x3] = ndgrid((1:n)/n); x = [x1(:) x2(:) x3(:)]';  % grid points

[…]
