### 2020-03-16 b9b4dd1
Document inputs for ifmm/test. Corrections.
