+++ ifmm/test/ie_circle.m
@@ -11,17 +11,27 @@
 %   - check multiply error/time
 %   - check solve error/time using GMRES
 %   - check PDE solve error by applying to known solution
+%
+% Inputs (defaults are used if not provided or set empty):
+%
+%   - N: number of discretization points (default: N = 16384)
+%   - OCC: tree occupancy parameter (default: OCC = 128)
+%   - P: number of proxy points (default: P = 64)
+%   - RANK_OR_TOL: local precision parameter (default: RANK_OR_TOL = 1e-12)
+%   - NEAR: near-field compression parameter (default: NEAR = 0)
+%   - STORE: storage parameter (default: STORE = 'A')
+%   - SYMM: symmetry parameter (default: SYMM = 'S')
 
 function ie_circle(N,occ,p,rank_or_tol,near,store,symm)
 
   % set default parameters
-  if nargin < 1 || isempty(N), N = 16384; end  % number of points
+  if nargin < 1 || isempty(N), N = 16384; end
   if nargin < 2 || isempty(occ), occ = 128; end
-  if nargin < 3 || isempty(p), p = 64; end  % number of proxy points
+  if nargin < 3 || isempty(p), p = 64; end
   if nargin < 4 || isempty(rank_or_tol), rank_or_tol = 1e-12; end
-  if nargin < 5 || isempty(near), near = 0; end  % no near-field compression
-  if nargin < 6 || isempty(store), store = 'a'; end  % store all interactions
-  if nargin < 7 || isempty(symm), symm = 's'; end  % symmetric
+  if nargin < 5 || isempty(near), near = 0; end
+  if nargin < 6 || isempty(store), store = 'a'; end
+  if nargin < 7 || isempty(symm), symm = 's'; end
 
   % initialize
   theta = (1:N)*2*pi/N; x = [cos(theta); sin(theta)];  % discretization points
