### 2020-03-16 5e2c772
Document inputs for hifie/test.
