### 2020-03-16 04b48a9
Document inputs for mf/test.
