+++ mf/test/fd_line2x.m
@@ -4,16 +4,24 @@
 % This is basically the same as FD_LINE1X but using a fourth-order approximation
 % with stencil radius two. Correspondingly, the nested dissection separators
 % have width two.
+%
+% Inputs (defaults are used if not provided or set empty):
+%
+%   - N: number of points + 1 in each dimension (default: N = 16384)
+%   - OCC: tree occupancy parameter (default: OCC = 64)
+%   - SYMM: symmetry parameter (default: SYMM = 'P')
+%   - DOITER: whether to run unpreconditioned CG (default: DOITER = 1)
+%   - DIAGMODE: diagonal extraction mode - 0: skip; 1: matrix unfolding; 2:
+%       sparse apply/solves (default: DIAGMODE = 0)
 
 function fd_line2x(n,occ,symm,doiter,diagmode)
 
   % set default parameters
-  if nargin < 1 || isempty(n), n = 16384; end  % number of points + 1
+  if nargin < 1 || isempty(n), n = 16384; end
   if nargin < 2 || isempty(occ), occ = 64; end
-  if nargin < 3 || isempty(symm), symm = 'p'; end  % positive definite
-  if nargin < 4 || isempty(doiter), doiter = 1; end  % unpreconditioned CG?
-  if nargin < 5 || isempty(diagmode), diagmode = 0; end  % diag extraction mode:
-  % 0 - skip; 1 - matrix unfolding; 2 - sparse apply/solves
+  if nargin < 3 || isempty(symm), symm = 'p'; end
+  if nargin < 4 || isempty(doiter), doiter = 1; end
+  if nargin < 5 || isempty(diagmode), diagmode = 0; end
 
   % initialize
   x = (1:n-1)/n; N = size(x,2);
