### 2020-03-16 3cd30c9
Document inputs for hifde/test.
