+++ hifde/test/fd_square5.m
@@ -17,18 +17,28 @@
 %
 %   - Repeat similarly with the right boundary, being careful to rescale the
 %     boundary rows of the matrix in order to maintain symmetry.
+%
+% Inputs (defaults are used if not provided or set empty):
+%
+%   - N: number of discretization points in each dimension (default: N = 128)
+%   - OCC: tree occupancy parameter (default: OCC = 8)
+%   - RANK_OR_TOL: local precision parameter (default: RANK_OR_TOL = 1e-9)
+%   - SKIP: skip parameter (default: SKIP = 2)
+%   - SYMM: symmetry parameter (default: SYMM = 'P')
+%   - DOITER: whether to run unpreconditioned CG (default: DOITER = 1)
+%   - DIAGMODE: diagonal extraction mode - 0: skip; 1: matrix unfolding; 2:
+%       sparse apply/solves (default: DIAGMODE = 0)
 
 function fd_square5(n,occ,rank_or_tol,skip,symm,doiter,diagmode)
 
   % set default parameters
-  if nargin < 1 || isempty(n), n = 128; end  % number of points + 1 in each dim
+  if nargin < 1 || isempty(n), n = 128; end
   if nargin < 2 || isempty(occ), occ = 8; end
   if nargin < 3 || isempty(rank_or_tol), rank_or_tol = 1e-9; end
   if nargin < 4 || isempty(skip), skip = 2; end
-  if nargin < 5 || isempty(symm), symm = 'p'; end  % positive definite
-  if nargin < 6 || isempty(doiter), doiter = 1; end  % unpreconditioned CG?
-  if nargin < 7 || isempty(diagmode), diagmode = 0; end  % diag extraction mode:
-  % 0 - skip; 1 - matrix unfolding; 2 - sparse apply/solves
+  if nargin < 5 || isempty(symm), symm = 'p'; end
+  if nargin < 6 || isempty(doiter), doiter = 1; end
+  if nargin < 7 || isempty(diagmode), diagmode = 0; end
 
   % initialize
   N = (n - 1)^2;  % total number of grid points
