+++ hifde/test/fd_cube3.m
@@ -2,19 +2,30 @@
 % Dirichlet boundary conditions.
 %
 % This is basically the 3D analogue of FD_SQUARE3.
+%
+% Inputs (defaults are used if not provided or set empty):
+%
+%   - N: number of discretization points in each dimension (default: N = 32)
+%   - K: wavenumber (default: K = 2*PI*4)
+%   - OCC: tree occupancy parameter (default: OCC = 4)
+%   - RANK_OR_TOL: local precision parameter (default: RANK_OR_TOL = 1e-6)
+%   - SKIP: skip parameter (default: SKIP = 2)
+%   - SYMM: symmetry parameter (default: SYMM = 'H')
+%   - DOITER: whether to run unpreconditioned CG (default: DOITER = 1)
+%   - DIAGMODE: diagonal extraction mode - 0: skip; 1: matrix unfolding; 2:
+%       sparse apply/solves (default: DIAGMODE = 0)
 
 function fd_cube3(n,k,occ,rank_or_tol,skip,symm,doiter,diagmode)
 
   % set default parameters
-  if nargin < 1 || isempty(n), n = 32; end  % number of points + 1 in each dim
-  if nargin < 2 || isempty(k), k = 2*pi*4; end  % wavenumber
+  if nargin < 1 || isempty(n), n = 32; end
+  if nargin < 2 || isempty(k), k = 2*pi*4; end
   if nargin < 3 || isempty(occ), occ = 4; end
   if nargin < 4 || isempty(rank_or_tol), rank_or_tol = 1e-6; end
   if nargin < 5 || isempty(skip), skip = 2; end
-  if nargin < 6 || isempty(symm), symm = 'h'; end  % positive definite
-  if nargin < 7 || isempty(doiter), doiter = 1; end  % unpreconditioned GMRES?
-  if nargin < 8 || isempty(diagmode), diagmode = 0; end  % diag extraction mode:
-  % 0 - skip; 1 - matrix unfolding; 2 - sparse apply/solves
+  if nargin < 6 || isempty(symm), symm = 'h'; end
+  if nargin < 7 || isempty(doiter), doiter = 1; end
+  if nargin < 8 || isempty(diagmode), diagmode = 0; end
 
   % initialize
   N = (n - 1)^3;  % total number of grid points
