### 2020-03-17 5546f74
Updates to README: more detail on ie_circle example; comment on dimension; spaces around em dashes; some other minor editing.
