### 2020-03-18 edfddf8
Remove double backticks (not sure why I had them to begin with); update paper; update bibliography; add emphasis in README example.
