+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'FLAM: Fast Linear Algebra in MATLAB'
+title: 'FLAM: Fast Linear Algebra in MATLAB - Algorithms for Hierarchical Matrices'
 tags:
   - hierarchical matrices
   - fast multipole method
@@ -16,37 +16,39 @@ authors:
 affiliations:
  - name: Center for Computational Mathematics, Flatiron Institute
    index: 1
-date: 6 November 2019
+date: 18 March 2020
 bibliography: paper.bib
 ---
 
 # Summary
 
-Many large matrices in science and engineering possess a special hierarchical low-rank structure that enables fast multiplication and inversion, among other fundamental operations. Such matrices commonly occur in physical problems, including the classical integral (IE) and differential equations (DE) of potential theory, as well as, for example, in kernel methods for high-dimensional statistics and machine learning. Owing to their ubiquity and practical importance, a wide variety of techniques for exploiting this structure have been developed, appearing in the literature under an assortment of related names such as the fast multiple method (FMM) [@greengard:1987:j-comput-phys], $\mathcal{H}$- and $\mathcal{H}^2$-matrices [@hackbusch:1999:computing], hierarchically semiseparable matrices [@xia:2010:numer-linear-algebra-appl], hierarchically block separable matrices [@gillman:2012:front-math-china], and hierarchical off-diagonal low-rank matrices [@ambikasaran:2013:j-sci-comput]. Many of these now also have accompanying open-source software packages: ``FMMLIB`` [@gimbutas:2015:commun-comput-phys], ``H2Lib`` (http://www.h2lib.org), ``STRUMPACK`` [@rouet:2016:acm-trans-math-softw; @ghysels:2016:siam-j-sci-comput], ``HODLRlib`` [@ambikasaran:2019:j-open-source-softw], ``hm-toolbox`` [@massei:2019:arxiv], etc.
+Many large matrices in science and engineering possess a special hierarchical low-rank structure that enables fast multiplication and inversion, among other fundamental operations. Such matrices commonly occur in physical problems, including the classical integral (IE) and differential equations (DE) of potential theory, as well as, for example, in kernel methods for high-dimensional statistics and machine learning. Owing to their ubiquity and practical importance, a wide variety of techniques for exploiting this structure have been developed, appearing in the literature under an assortment of related names such as the fast multiple method (FMM) [@greengard:1987:j-comput-phys], $\mathcal{H}$- and $\mathcal{H}^2$-matrices [@hackbusch:1999:computing], hierarchically semiseparable matrices [@xia:2010:numer-linear-algebra-appl], hierarchically block separable matrices [@gillman:2012:front-math-china], and hierarchical off-diagonal low-rank matrices [@ambikasaran:2013:j-sci-comput]. Many of these now also have accompanying open-source software packages: `FMMLIB` [@gimbutas:2015:commun-comput-phys], `H2Lib` (http://www.h2lib.org), `STRUMPACK` [@rouet:2016:acm-trans-math-softw; @ghysels:2016:siam-j-sci-comput], `HODLRlib` [@ambikasaran:2019:j-open-source-softw], `hm-toolbox` [@massei:2019:arxiv], etc.
 
[…]
