+++ paper/paper.bib
@@ -20,12 +20,14 @@
   doi = {10.21105/joss.01167}
 }
 
-@article{askham:2019:arxiv,
+@article{askham:2020:adv-comput-math,
   author = {Askham, T. and Rachh, M.},
   title = {{A boundary integral equation approach to computing eigenvalues of the Stokes operator}},
-  journal = {Preprint},
-  pages = {arXiv:1904.07351 [math.NA]},
-  year = {2019}
+  journal = {Adv. Comput. Math.},
+  volume = {46},
+  number = {2},
+  pages = {20},
+  year = {2020}
 }
 
 @article{corona:2017:j-comput-phys,
