+++ README.md
@@ -56,7 +56,7 @@ Download FLAM by typing, e.g.:
 git clone https://github.com/klho/FLAM
 ```
 
-This will create a local directory called ``FLAM`` containing the contents of the entire repository. Then enter the ``FLAM`` directory and launch MATLAB. This will automatically run FLAM's startup script and set up all required paths. Alternatively, if you launched MATLAB from a different directory or are using Octave instead, do this manually by running ``startup.m`` inside ``FLAM``.
+This will create a local directory called `FLAM` containing the contents of the entire repository. Then enter the `FLAM` directory and launch MATLAB. This will automatically run FLAM's startup script and set up all required paths. Alternatively, if you launched MATLAB from a different directory or are using Octave instead, do this manually by running `startup.m` inside `FLAM`.
 
 All FLAM functions should now be accessible. Test this by typing, e.g.:
 
@@ -64,7 +64,7 @@ All FLAM functions should now be accessible. Test this by typing, e.g.:
 help rskelf
 ```
 
-which should display the documentation for ``rskelf``:
+which should display the documentation for `rskelf`:
 
 ```
 RSKELF  Recursive skeletonization factorization.
@@ -74,16 +74,16 @@ RSKELF  Recursive skeletonization factorization.
 
 Other functions are similarly documented.
 
-All of the algorithms listed in the preamble are accompanied by extensive tests demonstrating their usage and performance. As an example, consider ``ie_circle.m`` inside ``FLAM/rskelf/test``, which solves a second-kind boundary IE on the unit circle. Briefly, the IE is derived from an underlying Laplace DE:
+All of the algorithms listed in the preamble are accompanied by extensive tests demonstrating their usage and performance. As an example, consider `ie_circle.m` inside `FLAM/rskelf/test`, which solves a second-kind boundary IE on the unit circle. Briefly, the IE is derived from an underlying Laplace DE:
 
     Δu(x) = 0     for  x ∊  Ω
      u(x) = f(x)  for  x ∊ ∂Ω
 
-on the unit disk `Ω` by writing the solution as a double-layer potential `u(x) = D[σ](x)` over an unknown surface density `σ`, where `D` is a convolution against the normal derivative of the Green's function. This representation satisfies the DE in the interior by construction; matching the boundary condition then gives the IE `Aσ = f`, where, in operator notation, `A = -1/2*I + D` for `D` now understood in the principal value sense. The matrix `A` is fully dense and thus can be very challenging to solve directly using conventional techniques. The purpose of this example is to demonstrate that in fact we can do so very efficiently, in linear time, by exploiting the inherent structure in `A`.
+on the unit disk `Ω` by writing the solution as a double-layer potential `u(x) = D[σ](x)` over an unknown surface density `σ`, where `D` is a convolution against the normal derivative of the Green's function. This representation satisfies the DE in the interior by construction; matching the boundary condition then gives the IE `Aσ = f`, where, in operator notation, `A = -1/2*I + D` for `D` now understood in the principal value sense. The matrix `A` is *fully dense* and thus can be very challenging to solve directly using conventional techniques. The purpose of this example is to demonstrate that in fact we can do so very efficiently, in linear time, by exploiting the inherent structure in `A`.
 
 As with all tests, `ie_circle` is actually written as a function, with parameters to set the number of discretization points, compression accuracy, symmetry properties, etc.; sensible defaults are provided for any omitted arguments. Please see the test file for more information.
 
-Running ``ie_circle`` then gives, e.g.:
+Running `ie_circle` then gives, e.g.:
 
 ```
 >> cd rskelf/test
@@ -109,13 +109,13 @@ pde solve err: 1.3504e-12
 
 The actual numbers may differ from run to run but should be roughly the same as those listed here. Let's now walk through this test in order and explain each output:
 
[…]
