### 2020-03-19 bb6a146
Update README title, add DOI for a reference.
