+++ paper/paper.bib
@@ -27,7 +27,8 @@
   volume = {46},
   number = {2},
   pages = {20},
-  year = {2020}
+  year = {2020},
+  doi = {10.1007/s10444-020-09774-2}
 }
 
 @article{corona:2017:j-comput-phys,
