### 2020-03-19 #1906 @klho
Please see the following changes:

0. I've added a subtitle to the paper so that it now reads "FLAM: Fast Linear Algebra in MATLAB - Algorithms for Hierarchical Matrices". A similar addition has been made to the README.

1. I've elaborated on the example in the README and tried to emphasize why the example is impressive. However, I'm a little at a loss on how to make it much more approachable: showing the discretized IE matrix would require a lot more background and notation, which I would really like to avoid; and even switching to a more "standard" PDE example would require some discussion of nested dissection, etc., which is arguably similarly specialized (at least for an introductory example). At any rate, let me know your thoughts on the current revision.

2. All test code arguments have now been documented.

3. Strictly speaking, there is no dimensional limitation in the code itself, but I have added statements to both the paper and the README.

4. Related software is cited in the opening paragraph. I've now also added another paragraph near the end summarizing some perhaps key distinctions.
