+++ index.md
@@ -1,5 +1,5 @@
-FLAM (Fast Linear Algebra in MATLAB)
-====================================
+FLAM (Fast Linear Algebra in MATLAB): Algorithms for Hierarchical Matrices
+==========================================================================
 
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1253581.svg)](https://doi.org/10.5281/zenodo.1253581)
 
@@ -26,7 +26,7 @@ It was originally intended as a research testbed for personal prototyping, thoug
 
 See [Algorithms](#algorithms) for details.
 
-FLAM has been written to emphasize readability and ease of use. This follows the choice of MATLAB as the computing platform, and we have preferred standard idiomatic MATLAB over more complicated constructions where possible. A key goal is for the codes to be readily understood at an introductory graduate level so that they may be broadly deployed, modified, and tested in an effort to make such fast matrix methods more accessible. Of course, this design may not yield the best performance; a more serious undertaking would re-implement the algorithms in a lower-level language---but this is beyond the scope of FLAM.
+FLAM has been written to emphasize readability and ease of use. This follows the choice of MATLAB as the computing platform, and we have preferred standard idiomatic MATLAB over more complicated constructions where possible. A key goal is for the codes to be readily understood at an introductory graduate level so that they may be broadly deployed, modified, and tested in an effort to make such fast matrix methods more accessible. Of course, this design may not yield the best performance; a more serious undertaking would re-implement the algorithms in a lower-level language -- but this is beyond the scope of FLAM.
 
 FLAM is freely available under the GNU GPLv3 and was last tested on MATLAB R2019a and Octave 4.2.2.
 
@@ -56,7 +56,7 @@ Download FLAM by typing, e.g.:
 git clone https://github.com/klho/FLAM
 ```
 
-This will create a local directory called ``FLAM`` containing the contents of the entire repository. Then enter the ``FLAM`` directory and launch MATLAB. This will automatically run FLAM's startup script and set up all required paths. Alternatively, if you launched MATLAB from a different directory or are using Octave instead, this will have to be done manually by running ``startup.m`` inside ``FLAM``.
+This will create a local directory called `FLAM` containing the contents of the entire repository. Then enter the `FLAM` directory and launch MATLAB. This will automatically run FLAM's startup script and set up all required paths. Alternatively, if you launched MATLAB from a different directory or are using Octave instead, do this manually by running `startup.m` inside `FLAM`.
 
 All FLAM functions should now be accessible. Test this by typing, e.g.:
 
@@ -64,19 +64,26 @@ All FLAM functions should now be accessible. Test this by typing, e.g.:
 help rskelf
 ```
 
-which should display the documentation for ``rskelf``:
+which should display the documentation for `rskelf`:
 
 ```
- RSKELF  Recursive skeletonization factorization.
+RSKELF  Recursive skeletonization factorization.
 
     The recursive skeletonization factorization  % ...
 ```
 
 Other functions are similarly documented.
 
-All of the algorithms listed in the preamble are accompanied by extensive tests demonstrating their usage and performance. As an example, consider ``ie_circle.m`` inside ``FLAM/rskelf/test``, which solves a second-kind boundary IE on the unit circle. This test is actually written as a function, with parameters to set the number of discretization points, compression accuracy, symmetry properties, etc.; sensible defaults are provided for any omitted arguments. Please see the test file for more information.
[…]
