### 2020-03-21 9c4f9e1
Add rskel_perm for "natural permutation" and use in rskel_mv -- keeps data always contiguous (but mostly will not be felt since vector usually fits fully into cache). Will also be useful later for other reasons.
