+++ rskel/rskel_mv.m
@@ -23,110 +23,108 @@ function Y = rskel_mv(F,X,trans)
   if trans == 't', Y = conj(rskel_mv(F,conj(X),'c')); return; end
 
   % initialize
-  M = F.M;
-  N = F.N;
   nlvl = F.nlvl;
-  rrem = true(M,1);
-  crem = true(N,1);
-  Z = cell(nlvl+1,1);  % successively compressed from upward pass
-  Y = cell(nlvl+1,1);  % successively refined from downward pass
+  if F.symm == 'n' && trans == 'c', p = F.q;
+  else,                             p = F.p;
+  end
+  if F.symm == 'n' && trans == 'n', q = F.q;
+  else,                             q = F.p;
+  end
+  np = length(p); prem = true(np,1);  % which "left"  indices remain?
+  nq = length(q); qrem = true(nq,1);  % which "right" indices remain?
+  P = zeros(nq,2);     % permutation for before/after current level
+  Z = cell(nlvl  ,1);  % successively compressed from   upward pass
+  Y = cell(nlvl+1,1);  % successively    refined from downward pass
 
   % upward sweep
-  Z{1} = X;
-  for lvl = 1:nlvl
-    prrem1 = cumsum(rrem);
-    pcrem1 = cumsum(crem);
+  P(q,1) = 1:nq;  % initial permutation
+  pf = 0;         % index for current level (to avoid data copy)
+  Z{1} = X(q,:);  % copy-permute from input
+  for lvl = 1:nlvl-1
+
+    % update permutation and copy-permute from lower level
     for i = F.lvpu(lvl)+1:F.lvpu(lvl+1)
-      rrem(F.U(i).rrd) = 0;
-      if F.symm == 'n', crem(F.U(i).crd) = 0;
-      else,             crem(F.U(i).rrd) = 0;
+      if F.symm == 'n' && trans == 'n', qrem(F.U(i).crd) = 0;
+      else,                             qrem(F.U(i).rrd) = 0;
       end
     end
-    prrem2 = cumsum(rrem);
-    pcrem2 = cumsum(crem);
-    if trans == 'n', Z{lvl+1} = Z{lvl}(pcrem1(find(crem)),:);
-    else,            Z{lvl+1} = Z{lvl}(prrem1(find(rrem)),:);
-    end
+    p1 =  pf + 1;
+    p2 = ~pf + 1;
+    pf = ~pf;
+    P(q(qrem(q)),p2) = 1:nnz(qrem);
+    Z{lvl+1}(P(qrem,p2),:) = Z{lvl}(P(qrem,p1),:);
 
     % apply interpolation operators
     for i = F.lvpu(lvl)+1:F.lvpu(lvl+1)
+      if F.symm == 'n' && trans == 'n'
+        rd = P(F.U(i).crd,p1);
+        sk = P(F.U(i).csk,p2);
+      else
+        rd = P(F.U(i).rrd,p1);
+        sk = P(F.U(i).rsk,p2);
+      end
       if F.symm == 'n'
-        if trans == 'n'
-          rd = pcrem1(F.U(i).crd);
-          sk = pcrem2(F.U(i).csk);
-          T = F.U(i).cT;
-        else
-          rd = prrem1(F.U(i).rrd);
-          sk = prrem2(F.U(i).rsk);
-          T = F.U(i).rT';
+        if trans == 'n', T = F.U(i).cT;
+        else,            T = F.U(i).rT';
         end
       elseif F.symm == 's'
[…]
