### 2020-03-21 f9b43aa
Do same for ifmm; simplify logic slightly.
