+++ ifmm/ifmm.m
@@ -159,8 +159,9 @@ function F = ifmm(A,rx,cx,occ,rank_or_tol,pxyfun,opts)
   %                    'D' for diagonal interactions and 'B' for others
   B = struct('is',e,'js',e,'ie',e,'je',e,'D',e,'Bo',e,'Bi',e);
   U = struct('rsk',e,'rrd',e,'csk',e,'crd',e,'rT',e,'cT',e);  % ID matrices
-  F = struct('M',M,'N',N,'nlvl',nlvl,'lvpb',zeros(1,nlvl+1),'lvpu', ...
+  F = struct('M',M,'N',N,'nlvl',nlvl,'lvpb',zeros(1,nlvl+2),'lvpu', ...
              zeros(1,nlvl+1),'B',B,'U',U,'store',opts.store,'symm',opts.symm);
+  [F.p,F.q] = ifmm_perm(t,opts.symm);
   nb = 0; mnb = nbox;  % number of B nodes and maximum capacity
   nu = 0; mnu = nbox;  % number of U nodes and maximum capacity
   rrem = true(M,1); crem = true(N,1);  % which row/cols remain?
@@ -303,8 +304,8 @@ function F = ifmm(A,rx,cx,occ,rank_or_tol,pxyfun,opts)
   end
 
   % loop over tree levels
-  nlvl = 1;              % offset for extra near-field level
-  for lvl = t.nlvl:-1:3  % no far field for top two levels
+  nlvl = 1;  % offset for extra near-field level
+  for lvl = t.nlvl:-1:1
     ts = tic;
     nlvl = nlvl + 1;
     nrrem1 = sum(rrem); ncrem1 = sum(crem);  % remaining row/cols at start
@@ -324,21 +325,29 @@ function F = ifmm(A,rx,cx,occ,rank_or_tol,pxyfun,opts)
       cnbr = [t.nodes(t.nodes(i).nbor).cxi];
 
       % compress row space
-      if isempty(pxyfun)
-        cfar = setdiff(find(crem),[cslf cnbr]);
-        K = A(rslf,cfar);
+      if lvl > 2
+        if isempty(pxyfun)
+          cfar = setdiff(find(crem),[cslf cnbr]);
+          K = A(rslf,cfar);
+        else
+          K = pxyfun('r',rx,cx,rslf,cnbr,l,t.nodes(i).ctr);
+        end
       else
-        K = pxyfun('r',rx,cx,rslf,cnbr,l,t.nodes(i).ctr);
+        K = zeros(length(rslf),0);
       end
       [rsk,rrd,rT] = id(K',rank_or_tol);
 
       % compress column space
       if opts.symm == 'n'
-        if isempty(pxyfun)
-          rfar = setdiff(find(rrem),[rslf rnbr]);
-          K = A(rfar,cslf);
+        if lvl > 2
+          if isempty(pxyfun)
+            rfar = setdiff(find(rrem),[rslf rnbr]);
+            K = A(rfar,cslf);
+          else
+            K = pxyfun('c',rx,cx,cslf,rnbr,l,t.nodes(i).ctr);
+          end
         else
-          K = pxyfun('c',rx,cx,cslf,rnbr,l,t.nodes(i).ctr);
+          K = zeros(0,length(cslf));
         end
         [csk,crd,cT] = id(K,rank_or_tol);
       else
@@ -377,49 +386,51 @@ function F = ifmm(A,rx,cx,occ,rank_or_tol,pxyfun,opts)
     F.lvpu(nlvl+1) = nu;
 
     % store far-field interactions
-    for i = t.lvp(lvl)+1:t.lvp(lvl+1)
[…]
