### 2020-03-21 c347684
Slightly cleaner.
