### 2020-03-21 1ebb47d
Use nnz instead of sum on logical arrays.
