+++ rskelf/rskelf.m
@@ -149,7 +149,7 @@ function F = rskelf(A,x,occ,rank_or_tol,pxyfun,opts)
   for lvl = t.nlvl:-1:1
     ts = tic;
     nlvl = nlvl + 1;
-    nrem1 = sum(rem);  % remaining points at start
+    nrem1 = nnz(rem);  % remaining points at start
     l = t.l(:,lvl);
 
     % pull up skeletons from children
@@ -239,7 +239,7 @@ function F = rskelf(A,x,occ,rank_or_tol,pxyfun,opts)
 
     % print summary
     if opts.verb
-      nrem2 = sum(rem);                              % remaining points at end
+      nrem2 = nnz(rem);                              % remaining points at end
       nblk = pblk(lvl) + t.lvp(lvl+1) - t.lvp(lvl);  % nonempty up to this level
       fprintf('%3d | %6d | %8d | %8d | %8.2f | %8.2f | %10.2e\n', ...
               lvl,nblk,nrem1,nrem2,nrem1/nblk,nrem2/nblk,te)
