+++ _config.yml
@@ -1,5 +1,5 @@
 title: FLAM
-description: Fast linear algebra in MATLAB: algorithms for hierarchical matrices
+description: "Fast linear algebra in MATLAB: algorithms for hierarchical matrices"
 google_analytics: 
 show_downloads: true
 theme: jekyll-theme-minimal
