### 2020-03-22 2caa360
Too enthusiastic: remove {rskel,ifmm}_perm since not really user-facing; replace with hypoct_perm. Simplify.
