### 2020-03-23 314c024
Extend rskel_xsp to allow output in "natural" order. Use in tests; can be dramatically faster. Fix fd_square, cov_*, and uls_* tests.
