+++ rskel/test/ie_circle.m
@@ -58,7 +58,7 @@ function ie_circle(N,occ,p,rank_or_tol,symm)
   fprintf('rskel_mv err/time: %10.4e / %10.4e (s)\n',err,t)
 
   % build extended sparsification
-  tic; A = rskel_xsp(F); t = toc;
+  tic; [A,p,q] = rskel_xsp(F); t = toc;
   w = whos('A'); mem = w.bytes/1e6;
   fprintf('rskel_xsp:\n')
   fprintf('  build time/mem: %10.4e (s) / %6.2f (MB)\n',t,mem);
@@ -70,7 +70,7 @@ function ie_circle(N,occ,p,rank_or_tol,symm)
     dolu = 1;
     A = A + tril(A,-1)';
   end
-  FA = struct('lu',dolu);
+  FA = struct('p',p,'q',q,'lu',dolu);
   tic
   if dolu, [FA.L,FA.U,FA.P] = lu(A);
   else,    [FA.L,FA.D,FA.P] = ldl(A);
@@ -151,7 +151,10 @@ end
 % sparse LU/LDL solve
 function Y = sv_(F,X,trans)
   N = size(X,1);
-  X = [X; zeros(size(F.L,1)-N,size(X,2))];
+  if trans == 'n', p = F.p; q = F.q;
+  else,            p = F.q; q = F.p;
+  end
+  X = [X(p,:); zeros(size(F.L,1)-N,size(X,2))];
   if F.lu
     if trans == 'n', Y = F.U \(F.L \(F.P *X));
     else,            Y = F.P'*(F.L'\(F.U'\X));
@@ -160,4 +163,5 @@ function Y = sv_(F,X,trans)
     Y = F.P*(F.L'\(F.D\(F.L\(F.P'*X))));
   end
   Y = Y(1:N,:);
+  Y(q,:) = Y;
 end

[…]
