### 2020-04-03 4dbbe5e
Can use nnz here.
