+++ paper/paper.bib
@@ -61,12 +61,15 @@
   doi = {10.1017/jfm.2017.150}
 }
 
-@article{feliu-faba:2018:arxiv,
+@article{feliu-faba:2020:commun-math-sci,
   author = {Feliu-Fab\`{a}, J. and Ho, K. L. and Ying, L.},
   title = {{Recursively preconditioned hierarchical interpolative factorization for elliptic partial differential equations}},
-  journal = {Preprint},
-  pages = {arXiv:1808.01364 [math.NA]},
-  year = {2018}
+  journal = {Commun. Math. Sci.},
+  volume = {18},
+  number = {1},
+  pages = {91-108},
+  year = {2020},
+  doi = {10.4310/CMS.2020.v18.n1.a4}
 }
 
 @article{ghysels:2016:siam-j-sci-comput,
