### 2020-04-05 #1906 @melissawm
Hello @vijaysm, would you be willing/available to review this software paper for JOSS? We do realize this may be a complicated time for all (you can read Arfon's message above about our current paused status). However, if you can let us know if you think you could spare some time to review this some time in the future (or not), please just mention it in this thread. Thank you in advance!
