### 2020-04-08 8ae6e7f
Store rT in rskel/ifmm differently so that T matrix from ID always has size sk x rd -- helps to make things somewhat easier.
