### 2020-04-13 25b070c
Slight simplification.
