+++ hifde/hifde3x.m
@@ -333,7 +333,6 @@ function F = hifde3x(A,x,occ,rank_or_tol,opts)
         % loop over blocks
         for i = 1:nblk
           slf = blk(i).xi;
-          sslf = sort(slf);
 
           % find neighbors
           [nbr,~] = find(A(:,slf));
@@ -342,7 +341,7 @@ function F = hifde3x(A,x,occ,rank_or_tol,opts)
             nbr = [nbr(:); nbrc(:)]';
           end
           nbr = unique(nbr);
-          nbr = nbr(~ismemb(nbr,sslf));
+          nbr = nbr(~ismemb(nbr,sort(slf)));
 
           % compress off-diagonal block
           K = spget(A,nbr,slf);
