+++ paper/paper.bib
@@ -222,12 +222,15 @@
   doi = {10.1137/060662253}
 }
 
-@article{massei:2019:arxiv,
+@article{massei:2020:siam-j-sci-comput,
   author = {Massei, S. and Robol, L. and Kressner, D.},
   title = {{hm-toolbox: Matlab software for HODLR and HSS matrices}},
-  journal = {Preprint},
-  pages = {arXiv:1909.07909 [math.NA]},
-  year = {2019}
+  journal = {SIAM J. Sci. Comput.},
+  volume = {42},
+  number = {2},
+  pages = {C43-C68},
+  year = {2020},
+  doi = {10.1137/19M1288048}
 }
 
 @article{minden:2016:multiscale-model-simul,
