+++ core/id.m
@@ -12,20 +12,40 @@
 %    [SK,RD,T] = ID(A,RANK_OR_TOL) sets K = RANK_OR_TOL if RANK_OR_TOL >= 1 and
 %    TOL = RANK_OR_TOL if RANK_OR_TOL < 1.
 %
-%    [SK,RD] = ID(A,...) skips the computation of the interpolation matrix T.
+%    [SK,RD,T] = ID(A,RANK_OR_TOL,TMAX) iteratively refines the approximation
+%    using rank-revealing QR column swaps until 1 <= MAX(ABS(T(:))) <= TMAX.
+%    Ignored if TMAX = INF (default: TMAX = 2). Refinement can improve numerical
+%    stability and reduce the output rank.
+%
+%    [SK,RD,T] = ID(A,RANK_OR_TOL,TMAX,RRQR_ITER) performs at most RRQR
+%    refinement iterations (default: RRQR_ITER = INF).
+%
+%    [SK,RD,T,NITER] = ID(A,RANK_OR_TOL,...) also returns the number NITER of
+%    RRQR iterations performed.
 %
 %    References:
 %
 %      H. Cheng, Z. Gimbutas, P.G. Martinsson, V. Rokhlin. On the compression of
-%        low rank matrices. SIAM J. Sci. Comput. 26(4): 1389-1404, 2005.
+%        low rank matrices. SIAM J. Sci. Comput. 26 (4): 1389-1404, 2005.
+%
+%      M. Gu, S.C. Eisenstat. Efficient algorithms for computing a strong rank-
+%        revealing QR factorization. SIAM J. Sci. Comput. 17 (4): 848-869, 1996.
 %
 %    See also QR.
 
-function [sk,rd,T] = id(A,rank_or_tol)
+function [sk,rd,T,niter] = id(A,rank_or_tol,Tmax,rrqr_iter)
+
+  % set default parameters
+  if nargin < 3 || isempty(Tmax), Tmax = 2; end
+  if nargin < 4 || isempty(rrqr_iter), rrqr_iter = Inf; end
 
   % check inputs
   assert(rank_or_tol >= 0,'FLAM:id:invalidRankOrTol', ...
          'Rank or tolerance must be nonnegative.')
+  assert(Tmax >= 1, 'FLAM:id:invalidTmax', ...
+         'Interpolation matrix entry bound must be >= 1.')
+  assert(rrqr_iter >= 0,'FLAM:id:invalidRRQRIter', ...
+         'Maximum number of RRQR iterations must be nonnegative.')
 
   % initialize
   [m,n] = size(A);
@@ -33,18 +53,147 @@ function [sk,rd,T] = id(A,rank_or_tol)
   % return if matrix is empty
   if isempty(A)
     sk = []; rd = 1:n;
-    if nargout > 2, T = zeros(0,n); end
+    T = zeros(0,n);
     return
   end
 
   % reduce row size if too rectangular
-  if m > 8*n, [Q,A] = qr(A,0); end
+  if m > 8*n, [~,A] = qr(A,0); end
 
   % compute ID
-  [~,R,P] = qr(A,0);
-  if rank_or_tol < 1, k = nnz(abs(diag(R)) > abs(R(1))*rank_or_tol);
-  else,               k = min(rank_or_tol,n);
+  [~,R,p] = qr(A,0);
+  if rank_or_tol < 1, tol = abs(R(1))*rank_or_tol;
+  else,               tol = 0;
   end
-  sk = P(1:k); rd = P(k+1:end);
-  if nargout > 2, T = R(1:k,1:k)\R(1:k,k+1:end); end
+  k = nnz(abs(diag(R)) > tol);
+  R = R(1:k,:);
[…]
