### 2020-05-21 9195d9d
Add ID-RRQR options to all algorithms.
