### 2020-05-21 7ee3358
Expose Tmax parameter in rskelf/tests.
