+++ rskelf/test/cov_circle2.m
@@ -8,23 +8,25 @@
 %   - OCC: tree occupancy parameter (default: OCC = 64)
 %   - P: square root of number of proxy points (default: P = 16)
 %   - RANK_OR_TOL: local precision parameter (default: RANK_OR_TOL = 1e-12)
+%   - TMAX: ID interpolation matrix entry bound (default: TMAX = 2)
 %   - SYMM: symmetry parameter (default: SYMM = 'P')
 %   - NOISE: nugget effect (default: NOISE = 1e-2)
 %   - SCALE: kernel length scale (default: SCALE = 100)
 %   - DIAGMODE: diagonal extraction mode - 0: skip; 1: matrix unfolding; 2:
 %       sparse apply/solves (default: DIAGMODE = 1)
 
-function cov_circle2(N,occ,p,rank_or_tol,symm,noise,scale,diagmode)
+function cov_circle2(N,occ,p,rank_or_tol,Tmax,symm,noise,scale,diagmode)
 
   % set default parameters
   if nargin < 1 || isempty(N), N = 16384; end
   if nargin < 2 || isempty(occ), occ = 64; end
   if nargin < 3 || isempty(p), p = 16; end
   if nargin < 4 || isempty(rank_or_tol), rank_or_tol = 1e-12; end
-  if nargin < 5 || isempty(symm), symm = 'p'; end
-  if nargin < 6 || isempty(noise), noise = 1e-2; end
-  if nargin < 7 || isempty(scale), scale = 100; end
-  if nargin < 8 || isempty(diagmode), diagmode = 1; end
+  if nargin < 5 || isempty(Tmax), Tmax = 2; end
+  if nargin < 6 || isempty(symm), symm = 'p'; end
+  if nargin < 7 || isempty(noise), noise = 1e-2; end
+  if nargin < 8 || isempty(scale), scale = 100; end
+  if nargin < 9 || isempty(diagmode), diagmode = 1; end
 
   % initialize
   theta = (1:N)*2*pi/N; x = [cos(theta); sin(theta)];  % row/col points
@@ -37,7 +39,7 @@ function cov_circle2(N,occ,p,rank_or_tol,symm,noise,scale,diagmode)
   % factor matrix
   Afun = @(i,j)Afun_(i,j,x,noise,scale);
   pxyfun = @(x,slf,nbr,l,ctr)pxyfun_(x,slf,nbr,l,ctr,proxy,scale);
-  opts = struct('symm',symm,'verb',1);
+  opts = struct('Tmax',Tmax,'symm',symm,'verb',1);
   tic; F = rskelf(Afun,x,occ,rank_or_tol,pxyfun,opts); t = toc;
   w = whos('F'); mem = w.bytes/1e6;
   fprintf('rskelf time/mem: %10.4e (s) / %6.2f (MB)\n',t,mem)

[…]
