### 2020-05-22 4bfd40f
Expose Tmax for rskel/tests.
