### 2020-05-22 00a82ca
Expose Tmax for ifmm/tests.
