+++ ifmm/test/mv_square1.m
@@ -17,10 +17,11 @@
 %   - OCC: tree occupancy parameter (default: OCC = 128)
 %   - P: number of proxy points (default: P = 64)
 %   - RANK_OR_TOL: local precision parameter (default: RANK_OR_TOL = 1e-9)
+%   - TMAX: ID interpolation matrix entry bound (default: TMAX = 2)
 %   - NEAR: near-field compression parameter (default: NEAR = 0)
 %   - STORE: storage parameter (default: STORE = 'N')
 
-function mv_square1(M,N,occ,p,rank_or_tol,near,store)
+function mv_square1(M,N,occ,p,rank_or_tol,Tmax,near,store)
 
   % set default parameters
   if nargin < 1 || isempty(M), M = 16384; end
@@ -28,8 +29,9 @@ function mv_square1(M,N,occ,p,rank_or_tol,near,store)
   if nargin < 3 || isempty(occ), occ = 128; end
   if nargin < 4 || isempty(p), p = 64; end
   if nargin < 5 || isempty(rank_or_tol), rank_or_tol = 1e-9; end
-  if nargin < 6 || isempty(near), near = 0; end
-  if nargin < 7 || isempty(store), store = 'n'; end
+  if nargin < 6 || isempty(Tmax), Tmax = 2; end
+  if nargin < 7 || isempty(near), near = 0; end
+  if nargin < 8 || isempty(store), store = 'n'; end
 
   % initialize
   rx = rand(2,M);                                              % row points
@@ -40,7 +42,7 @@ function mv_square1(M,N,occ,p,rank_or_tol,near,store)
   % compress matrix
   Afun = @(i,j)Afun_(i,j,rx,cx);
   pxyfun = @(rc,rx,cx,slf,nbr,l,ctr)pxyfun_(rc,rx,cx,slf,nbr,l,ctr,proxy);
-  opts = struct('near',near,'store',store,'verb',1);
+  opts = struct('Tmax',Tmax,'near',near,'store',store,'verb',1);
   tic; F = ifmm(Afun,rx,cx,occ,rank_or_tol,pxyfun,opts); t = toc;
   w = whos('F'); mem = w.bytes/1e6;
   fprintf('ifmm time/mem: %10.4e (s) / %6.2f (MB)\n',t,mem)
