### 2020-05-22 dad28b7
Expose Tmax for hifie/tests.
