+++ hifie/test/ie_cube2.m
@@ -8,18 +8,20 @@
 %   - OCC: tree occupancy parameter (default: OCC = 512)
 %   - P: number of proxy points (default: P = 512)
 %   - RANK_OR_TOL: local precision parameter (default: RANK_OR_TOL = 1e-3)
+%   - TMAX: ID interpolation matrix entry bound (default: TMAX = 2)
 %   - SKIP: skip parameter (default: SKIP = 0)
 %   - SYMM: symmetry parameter (default: SYMM = 'H')
 
-function ie_cube2(n,occ,p,rank_or_tol,skip,symm)
+function ie_cube2(n,occ,p,rank_or_tol,Tmax,skip,symm)
 
   % set default parameters
   if nargin < 1 || isempty(n), n = 32; end
   if nargin < 2 || isempty(occ), occ = 512; end
   if nargin < 3 || isempty(p), p = 512; end
   if nargin < 4 || isempty(rank_or_tol), rank_or_tol = 1e-3; end
-  if nargin < 5 || isempty(skip), skip = 0; end
-  if nargin < 6 || isempty(symm), symm = 'h'; end
+  if nargin < 5 || isempty(Tmax), Tmax = 2; end
+  if nargin < 6 || isempty(skip), skip = 0; end
+  if nargin < 7 || isempty(symm), symm = 'h'; end
 
   % initialize
   [x1,x2,x3] = ndgrid((1:n)/n); x = [x1(:) x2(:) x3(:)]';  % grid points
@@ -37,7 +39,7 @@ function ie_cube2(n,occ,p,rank_or_tol,skip,symm)
   % factor matrix
   Afun = @(i,j)Afun_(i,j,x,intgrl);
   pxyfun = @(x,slf,nbr,l,ctr)pxyfun_(x,slf,nbr,l,ctr,proxy);
-  opts = struct('skip',skip,'symm',symm,'verb',1);
+  opts = struct('Tmax',Tmax,'skip',skip,'symm',symm,'verb',1);
   tic; F = hifie3(Afun,x,occ,rank_or_tol,pxyfun,opts); t = toc;
   w = whos('F'); mem = w.bytes/1e6;
   fprintf('hifie3 time/mem: %10.4e (s) / %6.2f (MB)\n',t,mem)

[…]
