### 2020-05-22 e6fafd4
Expose Tmax for hifde/tests.
