+++ hifde/test/fd_cube2x.m
@@ -8,22 +8,24 @@
 %   - N: number of discretization points in each dimension (default: N = 32)
 %   - OCC: tree occupancy parameter (default: OCC = 64)
 %   - RANK_OR_TOL: local precision parameter (default: RANK_OR_TOL = 1e-6)
+%   - TMAX: ID interpolation matrix entry bound (default: TMAX = 2)
 %   - SKIP: skip parameter (default: SKIP = 2)
 %   - SYMM: symmetry parameter (default: SYMM = 'P')
 %   - DOITER: whether to run unpreconditioned CG (default: DOITER = 1)
 %   - DIAGMODE: diagonal extraction mode - 0: skip; 1: matrix unfolding; 2:
 %       sparse apply/solves (default: DIAGMODE = 0)
 
-function fd_cube2x(n,occ,rank_or_tol,skip,symm,doiter,diagmode)
+function fd_cube2x(n,occ,rank_or_tol,Tmax,skip,symm,doiter,diagmode)
 
   % set default parameters
   if nargin < 1 || isempty(n), n = 32; end
   if nargin < 2 || isempty(occ), occ = 64; end
   if nargin < 3 || isempty(rank_or_tol), rank_or_tol = 1e-6; end
-  if nargin < 4 || isempty(skip), skip = 2; end
-  if nargin < 5 || isempty(symm), symm = 'p'; end
-  if nargin < 6 || isempty(doiter), doiter = 1; end
-  if nargin < 7 || isempty(diagmode), diagmode = 0; end
+  if nargin < 4 || isempty(Tmax), Tmax = 2; end
+  if nargin < 5 || isempty(skip), skip = 2; end
+  if nargin < 6 || isempty(symm), symm = 'p'; end
+  if nargin < 7 || isempty(doiter), doiter = 1; end
+  if nargin < 8 || isempty(diagmode), diagmode = 0; end
 
   % initialize
   [x1,x2,x3] = ndgrid((1:n-1)/n); x = [x1(:) x2(:) x3(:)]';  % grid points
@@ -83,7 +85,7 @@ function fd_cube2x(n,occ,rank_or_tol,skip,symm,doiter,diagmode)
   clear idx Jl Sl Jr Sr Ju Su Jd Sd Jf Sf Jb Sb Im Sm I J S
 
   % factor matrix
-  opts = struct('skip',skip,'symm',symm,'verb',1);
+  opts = struct('Tmax',Tmax,'skip',skip,'symm',symm,'verb',1);
   tic; F = hifde3x(A,x,occ,rank_or_tol,opts); t = toc;
   w = whos('F'); mem = w.bytes/1e6;
   fprintf('hifde3x time/mem: %10.4e (s) / %6.2f (MB)\n',t,mem)

[…]
