### 2020-05-22 1bd3f1c
Generalize ID by allowing both rank/tolerance specification at once.
