### 2020-05-30 0b06a46
Add support for specifying a priori skeletons in ID; minor bug fix; correct doc.
