+++ core/id.m
@@ -1,17 +1,17 @@
 % ID  Interpolative decomposition.
 %
-%    [SK,RD,T] = ID(A,K) produces a rank-K approximation of A via the skeleton
-%    and redundant indices SK and RD, respectively, and an interpolation matrix
-%    T such that A(:,RD) = A(:,SK)*T + E, where LENGTH(SK) = K and NORM(E) is of
-%    order S(K+1) for S = SVD(A). If K > N = SIZE(A,2), then K is set equal to
-%    N.
+%    [SK,RD,T] = ID(A,KMAX) produces a structure-preserving approximation of A
+%    of rank K = MIN(KMAX,RANK(A)) via the skeleton and redundant indices SK and
+%    RD, respectively, and an interpolation matrix T such that
+%    A(:,RD) = A(:,SK)*T + E, where LENGTH(SK) = K and NORM(E) is of order
+%    S(KMAX+1) for S = SVD(A).
 %
 %    [SK,RD,T] = ID(A,TOL) produces a rank-adaptive approximation such that
 %    A(:,RD) = A(:,SK)*T + E, where NORM(E) is of order TOL*NORM(A).
 %
-%    [SK,RD,T] = ID(A,RANK_OR_TOL) sets K = RANK_OR_TOL if RANK_OR_TOL >= 1 and
-%    TOL = RANK_OR_TOL if RANK_OR_TOL < 1. More generally, both criteria can be
-%    specified simultaneously by setting RANK_OR_TOL = K + TOL, with the
+%    [SK,RD,T] = ID(A,RANK_OR_TOL) sets KMAX = RANK_OR_TOL if RANK_OR_TOL >= 1
+%    and TOL = RANK_OR_TOL if RANK_OR_TOL < 1. More generally, both criteria can
+%    be specified simultaneously by letting RANK_OR_TOL = KMAX + TOL, with the
 %    approximation terminating as soon as either is reached.
 %
 %    [SK,RD,T] = ID(A,RANK_OR_TOL,TMAX) iteratively refines the approximation
@@ -19,11 +19,18 @@
 %    Ignored if TMAX = INF (default: TMAX = 2). Refinement can improve numerical
 %    stability and reduce the output rank.
 %
-%    [SK,RD,T] = ID(A,RANK_OR_TOL,TMAX,RRQR_ITER) performs at most RRQR
+%    [SK,RD,T] = ID(A,RANK_OR_TOL,TMAX,RRQR_ITER) performs at most RRQR_ITER
 %    refinement iterations (default: RRQR_ITER = INF).
 %
 %    [SK,RD,T,NITER] = ID(A,RANK_OR_TOL,...) also returns the number NITER of
-%    RRQR iterations performed.
+%    refinement iterations performed.
+%
+%    [SK,RD,T,...] = ID(A,RANK_OR_TOL,TMAX,RRQR_ITER,FIXED) forces the indices
+%    in FIXED to be included in SK (default: FIXED = []). An ID is computed on
+%    the residual free columns then reconstituted along with the fixed ones. The
+%    parameters TMAX and RRQR_ITER apply only to this residual ID. The maximum
+%    allowable rank KMAX = FLOOR(RANK_OR_TOL) is imposed on the full ID, with
+%    the exception that if LENGTH(FIXED) >= KMAX, then SK = FIXED.
 %
 %    References:
 %
@@ -35,11 +42,12 @@
 %
 %    See also QR.
 
-function [sk,rd,T,niter] = id(A,rank_or_tol,Tmax,rrqr_iter)
+function [sk,rd,T,niter] = id(A,rank_or_tol,Tmax,rrqr_iter,fixed)
 
   % set default parameters
[…]
