### 2020-05-31 8e789a4
Compatibility for sometimes weird MATLAB/Octave treatment of scalars vs. vectors vs. matrices, logical assignment/indexing.
