+++ ifmm/ifmm_mv.m
@@ -48,8 +48,8 @@ function Y = ifmm_mv(F,X,A,trans)
 
     % update permutation and copy-permute from lower level
     for i = F.lvpu(lvl)+1:F.lvpu(lvl+1)
-      if F.symm == 'n' && trans == 'n', qrem(F.U(i).crd) = 0;
-      else,                             qrem(F.U(i).rrd) = 0;
+      if F.symm == 'n' && trans == 'n', qrem(F.U(i).crd) = false;
+      else,                             qrem(F.U(i).rrd) = false;
       end
     end
     p1 =  pf + 1;
@@ -82,7 +82,7 @@ function Y = ifmm_mv(F,X,A,trans)
   end
 
   % downward sweep
-  prem(:) = 0;
+  prem(:) = false;
   P = zeros(np,2);  % reset permutations
   Q = zeros(nq,1);  % permutation for data from upward sweep
   pf = 0;
@@ -93,11 +93,11 @@ function Y = ifmm_mv(F,X,A,trans)
     % update permutation and copy-permute from higher level
     r = p(prem(p));
     for i = F.lvpu(lvl)+1:F.lvpu(lvl+1)
-      if F.symm == 'n' && trans == 'c', prem(F.U(i).crd) = 1;
-      else,                             prem(F.U(i).rrd) = 1;
+      if F.symm == 'n' && trans == 'c', prem(F.U(i).crd) = true;
+      else,                             prem(F.U(i).rrd) = true;
       end
-      if F.symm == 'n' && trans == 'n', qrem(F.U(i).crd) = 1;
-      else,                             qrem(F.U(i).rrd) = 1;
+      if F.symm == 'n' && trans == 'n', qrem(F.U(i).crd) = true;
+      else,                             qrem(F.U(i).rrd) = true;
       end
     end
     p1 =  pf + 1;
