+++ hifde/hifde3x.m
@@ -212,7 +212,7 @@ function F = hifde3x(A,x,occ,rank_or_tol,opts)
 
           % move on if no compression
           if isempty(rd), continue; end
-          rem(slf(rd)) = 0;
+          rem(slf(rd)) = false;
 
           % store data
           nb = nb + 1;
@@ -271,8 +271,8 @@ function F = hifde3x(A,x,occ,rank_or_tol,opts)
             dx = abs(c(1,:)' - c(1,:))/l(1);  % scaled distance between boxes
             dy = abs(c(2,:)' - c(2,:))/l(2);
             dz = abs(c(3,:)' - c(3,:))/l(3);
-            dist = round(dx + dy + dz);              % convert to integer
-            if any(dist(:) == 2), keep(ip) = 1; end  % diagonal -> dist = 2
+            dist = round(dx + dy + dz);                 % convert to integer
+            if any(dist(:) == 2), keep(ip) = true; end  % diagonal -> dist = 2
           end
           p = p(keep);
         end
@@ -362,7 +362,7 @@ function F = hifde3x(A,x,occ,rank_or_tol,opts)
 
           % move on if no compression
           if isempty(rd), continue; end
-          rem(slf(rd)) = 0;
+          rem(slf(rd)) = false;
 
           % store data
           nb = nb + 1;
