+++ mf/mfx.m
@@ -186,7 +186,7 @@ function F = mfx(A,x,occ,opts)
 
       % move on if no compression
       if isempty(rd), continue; end
-      rem(slf(rd)) = 0;
+      rem(slf(rd)) = false;
 
       % compute factors
       K = spget(A,slf,slf);
