+++ hifde/hifde2.m
@@ -121,7 +121,7 @@ function F = hifde2(A,n,occ,rank_or_tol,opts)
 
           % move on if no compression
           if isempty(rd), continue; end
-          rem(slf(rd)) = 0;
+          rem(slf(rd)) = false;
 
           % store data
           nb = nb + 1;
@@ -183,7 +183,7 @@ function F = hifde2(A,n,occ,rank_or_tol,opts)
 
           % move on if no compression
           if isempty(rd), continue; end
-          rem(slf(rd)) = 0;
+          rem(slf(rd)) = false;
 
           % store data
           nb = nb + 1;
