### 2020-06-13 19baa70
Set niter = -1 in snorm if no convergence (like in lsedc); simplify. Improve lsedc doc.
