### 2020-06-22 #1906 @vijaysm
@melissawm @arfon I have reviewed the software, repository, verified functionality and the accompanying article. I recommend the acceptance of the paper and publication in JOSS. I think this will be a useful toolbox for tackling matrices with a hierarchical structure as the authors have clearly pointed out applications and research areas where they are already being used.

Some suggestion and comments for @klho to improve the submission.

- It might be interesting to have a comparative figure showcasing performance of the hierarchical solvers in FLAM against standard generalized LU solvers for dense matrices or PCG with ILU preconditioners for sparse matrices in MATLAB. This will give the readers a better sense of performance gains that can be achieved with FLAM. The README file in the repository does have some performance numbers for both time and memory complexity, and can be highlighted in the paper. I also can see that some of the references do help in this regard as well.
- Please mention minimum required versions of MATLAB and Octave to use FLAM in the article and the README file in the repository
- Fix the typo in the article: FMM is Fast **Multipole** Method and not **Multiple*. Capitalize appropriately as well.
