### 2020-06-23 0e7d8a6
Fix typo in paper, add brief numerical highlights.
