### 2020-06-23 #1906 @klho
@vijaysm Thanks for the feedback; please see my responses below:

1. I've added a brief paragraph highlighting some numerical examples. I'm hesitant to put much more in the paper since I'd like to keep it light and avoid having to give too much background explanation. (Many) more details can be found in the cited papers.

2. This is quite reasonable but unfortunately might not doable. I don't have access to old versions of MATLAB to test. While I can download old versions of Octave, this is very time-consuming and ultimately may not be very informative. What I can say is that only "core" features of MATLAB/Octave are used, so compatibility should be quite stable (up to harmless warnings, e.g., for broadcasting), at least since 2014 when this project was initiated (and likely much further). But I have no way of making that statement precise, which is why I'd prefer to leave it off altogether. Please let me know your thoughts and how strongly you feel about this.

3. Thanks for catching the typo! However, I don't agree with the capitalization and so have not changed that; please see https://en.wikipedia.org/wiki/Wikipedia:Manual_of_Style/Capital_letters#Expanded_forms_of_abbreviations.
