+++ paper/paper.md
@@ -42,7 +42,7 @@ Similarly, for sparse DE-like matrices, we have:
 
 Most of these have previously been published though some are perhaps new (if but straightforward modifications or extensions of existing ones).
 
-Each algorithm comes with extensive tests demonstrating its usage and performance. For instance, `rskelf/tests/ie_circle.m` solves a second-kind Laplace boundary IE on the unit circle. At default settings, the problem is characterized by a *fully dense* square matrix of size $N = 16384$, which takes about 58 s to factor using MATLAB's `lu` at a storage cost of 2 GB. In contrast, `rskelf` requires just 0.7 s to compute an LU-like approximation to $10^{-12}$ precision, a speedup of about 80$\times$; once computed, the factorization needs only 9 MB in storage and can be used to execute solves in 0.04 s. Another exemplary test case is `hifde/tests/fd_square2.m`, which considers a Poisson DE on a very rough and high-contrast background field. Standard preconditioners such as `ichol` do not deal well with such severe ill-conditioning, but `hifde` remains highly effective, producing preconditioners that converge in just a handful of iterations. For more detailed performance analyses, we refer the reader to the cited literature above.
+Each algorithm comes with extensive tests demonstrating its usage and performance. For instance, `rskelf/tests/ie_circle.m` solves a second-kind Laplace boundary IE on the unit circle. At default settings, the problem is characterized by a *fully dense* square matrix of size $N = 16384$, which takes about 60 s to factor using MATLAB's `lu` at a storage cost of 2 GB. In contrast, `rskelf` requires just 0.7 s to compute an LU-like approximation to $10^{-12}$ precision, a speedup of about 80$\times$; once computed, the factorization needs only 9 MB to store and can be used to execute solves in 0.04 s. Another exemplary test case is `hifde/tests/fd_square2.m`, which considers a Poisson DE on a very rough and high-contrast background field. Standard preconditioners such as `ichol` do not deal well with such severe ill-conditioning, but `hifde` remains highly effective, producing preconditioners that converge in just a handful of iterations. For more detailed performance analyses, we refer the reader to the cited literature above.
 
 The provided functionality is somewhat similar to that offered by other software packages, but `FLAM` implements certain advanced methods like `hifie` and `hifde`, which to our knowledge are not publicly available elsewhere (though a close relative of `hifde` can be found in `STRUMPACK`). Furthermore, `FLAM` leverages a different core framework that we believe is considerably simpler and particularly well-suited to MATLAB's concise and readable style.
 
