### 2020-06-26 37d91de
Print out imaginary part of logdet also.
