+++ mf/test/fd_square5.m
@@ -120,10 +120,10 @@ function fd_square5(n,occ,symm,doiter,diagmode)
   fprintf('  precon/unprecon iter: %d / %d\n',piter,iter)
 
   % compute log-determinant
-  tic
-  ld = mf_logdet(F);
-  t = toc;
-  fprintf('mf_logdet: %22.16e / %10.4e (s)\n',ld,t)
+  tic; ld = mf_logdet(F); t = toc;
+  fprintf('mf_logdet:\n')
+  fprintf('  real/imag: %22.16e / %22.16e\n',real(ld),imag(ld))
+  fprintf('  time: %10.4e (s)\n',t)
 
   if diagmode > 0
     % prepare for diagonal extraction
