### 2020-07-03 #1906 @vijaysm
I apologize. I thought I had replied to @klho's comments.

@klho Thanks for updating the paper.

1. I agree that the citations have more details but I just wanted you to highlight some clear advantages of the hierarchical solvers in FLAM to give the reader a better idea before diving deeper into using the software implementation.

2. For the minimum version, you can go ahead and mention the MATLAB/Octave versions that you have tested so far, even if these are the latest ones. It just eliminates confusion from a user standpoint on which version would work for FLAM.

3. This is fine. It was more of an aesthetic suggestion but as long as you are consistent throughout the paper and in README, I think we can leave it as is.

@arfon I am satisfied with the author response. After a minor update to include details for [2] in the README, you can go ahead and accept the paper. Thanks.
