### 2020-07-03 #1906 @klho
@vijaysm There is already a statement on the last tested version in the README. Is that sufficient?
