### 2020-07-03 #1906 @vijaysm
From README.md in the repository:

> FLAM is freely available under the GNU GPLv3 and was last tested on MATLAB R2019a and Octave 4.2.2.

Excellent. Yes, this will suffice.
